backend/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pytz==2024.1
requests==2.31.0
openpyxl==3.1.2
aiohttp==3.9.1
geopandas==0.14.1
shapely==2.0.2
pyarrow==12.0.0
duckdb==0.9.2
lightgbm==3.1.1
catboost==1.2.2
joblib==1.3.2
pandas==2.1.4
numpy==1.24.3
google-cloud-storage==2.10.0
reportlab==4.0.8
Pillow==10.1.0
httpx==0.25.2
